Comment out problematic code for debugging mock build issue
