+++ runtime/common/src/lib.rs
@@ -41,6 +41,8 @@ pub mod precompiles;
 #[cfg(test)]
 mod mock;
 
+#[cfg(test)]
+mod tests;
 /// Simple fee calculator that requires payment in a single fungible at a fixed rate.
 ///
 /// - The `FixedRate` constant should be the concrete fungible ID and the amount of it
